tidy up spacing, move utils out of getters
